Modularize web server
* Split Config handlers out of core.py
* Optimize imports
